SprightBlight V0.02
We have now multiple screens (with some problems, but text and formatting can be resized) that the user can cycle though. This is the fundamentals to some of the interact gestures I'll be using.

I will be keeping the scope of this game relavtively small for now, just
cause I want the ability to have a workable game, since working with raylib is not the easiest task