SprightBlight V0.03
My build of the color palette is essential for a lot of the menu and option functionalities. While color may matter significantly less, these are a few great options in custom options that differs from the stock of raylib.

I will keep the stock palette, I just want more options that are reflected of the game