SprightBlight V0.08
Audio is starting to get handled properly. I've added a very important aspect, which is different music option for gameplay. This is not utilized yet in gameplay, however, the code schema is there and ready for it to be built out.

This has been honestly one of my favorite projects to work on so far. With so much modularity that makes this far better than working with Unity.

Coding this in C is honestly not that big a deal, and this flexibility is what makes working in this language so much fun!