SprightBlight V0.09
Music is now relative to the correct screens, allowing me to now build out the audio according.

This is great, since now I can focus on audio development whenever that is applicable.

What will increment this program by a version is a jump mechanic, as that next level of interactivity is what the game needs.

So far, this game is working as intended. Meaning that we will get close to a basic build of the game