SprightBlight V0.09
More of the necessary functions for the redo of this code. I understand this will not compile at all (some issues already with multiple calls to player), for there is a lot to refactor into said code. It should hopefully be as simple as deleting the older logic, and including any crucial pieces of code